64-bit changes
SCCS-vsn: 8.2

+++ sys/netiso/if_eon.c
@@ -30,7 +30,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)if_eon.c	8.1 (Berkeley) 06/10/93
+ *	@(#)if_eon.c	8.2 (Berkeley) 01/09/95
  */
 
 /***********************************************************
@@ -178,7 +178,7 @@ eonattach()
  */
 eonioctl(ifp, cmd, data)
 	register struct ifnet *ifp;
-	int cmd;
+	u_long cmd;
 	register caddr_t data;
 {
 	int s = splimp();
